+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - Added `\Wizaplace\SDK\Catalog\ProductAttribute::getValues`
 - Deprecated `\Wizaplace\SDK\Catalog\ProductAttribute::getValueIds`
 - Deprecated `\Wizaplace\SDK\Catalog\ProductAttribute::getValue`
+- Added `\Wizaplace\SDK\Discussion\DiscussionService::startDiscussionWithVendor`
+- Added `\Wizaplace\SDK\Discussion\DiscussionService::startDiscussionFromDeclinationId`
 
 ### Bugfixes